Replace the PolyDatabase by a ByteSlice Database
